Moved mox shared classes to avoid circular dependencies
